requirements.txt
FastAPI[all]==0.95.1
pydantic==1.10.7
uvicorn
pyjwt
requests
python-dotenv
pytest==7.3.1
coverage==7.2.3
cryptography
python-jose
mangum